Refactoring - common.mktemp()/rmtemp() are removed; use modified XDWTemp class instead.

+++ xdwlib/common.py
@@ -13,6 +13,7 @@ WARRANTIES OF TITLE, MERCHANTABILITY, AGAINST INFRINGEMENT, AND FITNESS
 FOR A PARTICULAR PURPOSE.
 """
 
+import sys
 import os
 import re
 from tempfile import mkstemp, mkdtemp
@@ -35,7 +36,7 @@ __all__ = (
         "mm2in", "in2mm", "mm2px", "px2mm",
         "environ", "get_viewer",
         "inner_attribute_name", "outer_attribute_name",
-        "adjust_path", "cp", "uc", "derivative_path", "mktemp", "rmtemp",
+        "adjust_path", "cp", "uc", "derivative_path",
         "joinf", "flagvalue", "typevalue", "makevalue", "scale", "unpack",
         "XDWTemp",
         )
@@ -241,20 +242,6 @@ def derivative_path(path):
     return derivative
 
 
-def mktemp(suffix=".xdw", prefix="", nofile=False):
-    fd, temp = mkstemp(suffix=suffix, prefix=prefix, dir=mkdtemp())
-    os.close(fd)
-    if nofile:
-        os.remove(temp)  # Directory is not removed.
-    return temp
-
-
-def rmtemp(path):
-    if os.path.exists(path):
-        os.remove(path)
-    os.rmdir(os.path.split(path)[0])
-
-
 def flagvalue(table, value, store=True):
     """Sum up flag values according to XDWConst table."""
     if store and isinstance(value, (int, float)):
@@ -331,37 +318,27 @@ def unpack(s):
 
 class XDWTemp(object):
 
-    """Temporary XDW file with optional single blank page."""
+    """Prepare a reusable pathname for a temporary file."""
 
-    def __init__(self, suffix=".xdw", dir=None, blank_page=False):
-        args = [suffix, "", dir] if dir else [suffix]
-        self.fd, self.path = mkstemp(*args)
-        if blank_page:
-            os.write(self.fd, BLANKPAGE)
+    def __init__(self, suffix=".xdw", prefix=""):
+        fd, path = mkstemp(suffix=suffix, prefix=prefix, dir=mkdtemp())
+        os.close(fd)
+        os.remove(path)  # Directory is not removed.
+        self.path = path
+        self.dir = os.path.split(path)[0]
+
+    def close(self):
+        if os.path.exists(self.path):
+            os.remove(self.path)
+        try:
+            os.rmdir(self.dir)
+        except Exception as e:
+            sys.stderr.write("""\
+{0}:xdwlib:can't delete temporary directory '{1}'\n""".format(
+datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"), self.dir))
 
     def __enter__(self):
         return self
 
-    def __exit__(self):
+    def __exit__(self, exc_type, exc_val, exc_tb):
         self.close()
-
-    def __del__(self):
-        try:
-            os.fdopen(self.fd).close()
-        except:
-            pass
-        os.remove(self.path)
-
-    def close(self):
-        os.fdopen(self.fd).close()
-
-    def seek(self, pos, how=0):
-        """Change read/write pointer in the temporary file.
-
-        pos     position in stream by bytes
-        how     base; 0=top, 1=current, 2=bottom
-        """
-        os.lseek(self.fd, pos, how)
-
-    def read(self, size):
-        return os.read(self.fd, size)
